Created concern for status of user, added show action
