+++ fbclone/app/controllers/application_controller.rb
@@ -1,3 +1,4 @@
 class ApplicationController < ActionController::Base
 	include DeviseWhitelist
+	include UserStatus
 end
