+++ fbclone/app/controllers/posts_controller.rb
@@ -1,4 +1,9 @@
 class PostsController < ApplicationController
+	before_action :is_not_current_user?, only: [:index]
+	
   def index
+  end
+
+  def show
   end
 end
