+++ fbclone/config/routes.rb
@@ -2,6 +2,6 @@ Rails.application.routes.draw do
   devise_for :users
 
   resource :profile, only: [:show, :edit], path: '/:username' do 
-    resources :posts, only: [:index]
+    resources :posts, only: [:index, :show]
   end
 end
